Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.0